modify resolution 640x480 to 320x480

+++ app/src/main/java/com/marvoto/fat/MeasureDepth.java
@@ -26,7 +26,7 @@ public final class MeasureDepth{
      * -3     图像不清晰
      * >0     测量脂肪厚度
      */
-    public static native  int DrawBitmap(Bitmap var0, byte[] var1);
+    public static native  float  DrawBitmap(Bitmap var0, byte[] var1);
 
     /**
      * desc   Set parameter from fpga
